requirements.txt
langchain-mistralai
langchain
mistralai 
tiktoken 
rapidocr-onnxruntime 
python-dotenv 
langchain-community